a6bc110 patched parameter type bug that was causing failure on 32 bit (Dion Ahmetaj)

+++ src/cryptonote_core/cryptonote_core.cpp
@@ -616,7 +616,7 @@ namespace cryptonote
     return true;
   }
   //-----------------------------------------------------------------------------------------------
-  std::pair<uint64_t, uint64_t> core::get_coinbase_tx_sum(const uint64_t start_offset, const uint64_t count)
+  std::pair<uint64_t, uint64_t> core::get_coinbase_tx_sum(const uint64_t start_offset, const size_t count)
   {
     std::list<block> blocks;
     std::list<transaction> txs;
